Moved the networking directory + Added a badge for the github acionts workflow
