+++ app/src/main/java/nick/mirosh/newsapp/data/repository/NewsRemoteDataSource.kt
@@ -1,7 +1,7 @@
 package nick.mirosh.newsapp.data.repository
 
 import nick.mirosh.newsapp.data.models.ArticleDTO
-import nick.mirosh.newsapp.networking.NewsService
+import nick.mirosh.newsapp.data.networking.NewsService
 import javax.inject.Inject
 
 class NewsRemoteDataSource @Inject constructor(private val newsService: NewsService) {
